# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Online Exam Cheating Detection API
  version: 0.1.0
  description: REST API for sessions, ingestion, incidents, and reviews.
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UUID:
      type: string
      format: uuid
...
    SessionStatus:
      type: string
      enum: [ACTIVE, ENDED, ABORTED]
...
    Session:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        user_id: { $ref: '#/components/schemas/UUID' }
        exam_id: { $ref: '#/components/schemas/UUID' }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
        status: { $ref: '#/components/schemas/SessionStatus' }
...
paths:
...
  /sessions/start:
    post:
      summary: Start exam session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [exam_id]
              properties:
                exam_id: { $ref: '#/components/schemas/UUID' }
      responses:
        '200':
          description: Session started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
  /sessions/{id}/end:
    post:
      summary: End exam session
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/UUID' }
      responses:
        '200': { description: Session ended }
...